combined unit tests for legacy and refactored code into one class
